Extract sort mode to separate class
