+++ src/main/java/com/cubefury/vendingmachine/blocks/gui/MTEVendingMachineGui.java
@@ -1,8 +1,5 @@
 package com.cubefury.vendingmachine.blocks.gui;
 
-import static com.cubefury.vendingmachine.gui.GuiTextures.SORT_ALPHABET;
-import static com.cubefury.vendingmachine.gui.GuiTextures.SORT_SMART;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -18,8 +15,6 @@ import net.minecraft.item.ItemStack;
 import com.cleanroommc.modularui.api.drawable.IKey;
 import com.cleanroommc.modularui.api.widget.IWidget;
 import com.cleanroommc.modularui.drawable.DynamicDrawable;
-import com.cleanroommc.modularui.drawable.Icon;
-import com.cleanroommc.modularui.drawable.UITexture;
 import com.cleanroommc.modularui.factory.PosGuiData;
 import com.cleanroommc.modularui.screen.ModularPanel;
 import com.cleanroommc.modularui.screen.RichTooltip;
@@ -95,29 +90,6 @@ public class MTEVendingMachineGui extends MTEMultiBlockBaseGui {
 
     private static final int COIN_COLUMN_WIDTH = 40;
     private static final int COIN_COLUMN_ROW_COUNT = 4;
-
-    public enum SortMode {
-
-        SMART("smart", SORT_SMART),
-        ALPHABET("alphabet", SORT_ALPHABET);
-
-        private String mode;
-        private Icon texture;
-
-        SortMode(String mode, UITexture texture) {
-            this.mode = mode;
-            this.texture = texture.asIcon();
-        }
-
-        public String getLocalizedName() {
-            return IKey.lang("vendingmachine.gui.display_sort_" + this.mode)
-                .toString();
-        }
-
-        public Icon getTexture() {
-            return this.texture;
-        }
-    }
 
     public MTEVendingMachineGui(MTEVendingMachine base) {
         super(base);
